Create most_severe_sentence_and_charge_spans (Recidiviz/recidiviz-data#44808)
## Description of the change

This PR creates a new view called
`most_severe_sentence_and_charge_spans` that contains two ids, that are
relevant for a person at each point in time:
1. `sentences_and_charges_id` for the record in
`sentence_sessions.sentences_and_charges` that represents the most
severe sentence and charge
2. `session_id` for the session in `sessions.compartment_sessions` that
they are in at that point

This allows us to easily join back to these tables to determine most
severe sentence/charge attributes and session attributes associated with
each person over time.

This is view was created by adding a second array to
`overlapping_sentence_serving_periods` that represents the set of
sentences that a person is not currently serving, but was previously
serving within the sentence_inferred_group_id that they are currently
in. This allows us to capture all relevant sentences when calculating
the most severe.

I created tests for both `overlapping_sentence_serving_periods` and
`most_severe_sentence_and_charge_spans` as part of this PR to confirm
that logic is working as expected.

I have a first pass demo set of PS visuals off of this new infra in a
notebook
[here](https://colab.research.google.com/drive/1Kx_oHDXiUpJfOwTlOaQYgM1dSaYDB2CB?usp=sharing).
The queries are simplified significantly following the creation of this
new view.

Lastly, I made an semi-related change to the `gen_cohort_status_df`
function in `cohort_methods` file that allows us to past a list of
columns to be considered as part of the status attributes (as opposed so
a single column). This is useful for the chart that looks at outcome
over time based on both compartment and charge severity.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 23ddaa2b13ac245c36dd22db582297a17933f3b4