[bq] add `failure_mode` arg to the view update (Recidiviz/recidiviz-data#45654)
## Description of the change

this PR adds `failure_mode` to the view update. as described in the
[project](https://www.notion.so/recidiviz/Declarative-BigQueryView-Schemas-22d7889f4d19801a91bdea6a40af1c58?source=copy_linkRecidiviz/recidiviz-data#2317889f4d1980cf99bcf6e22633aad5)
docs, this defaults to `EXHAUSTIVE` for all places except for sandbox
loads. adds the `--failure-mode` parameter to the `load_view_to_sandbox`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

n/a

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b91227641726209019adf2b0c3d16fd047e6393a