[US_AZ] Ingest StatePersonAddressPeriod (Recidiviz/recidiviz-data#45695)
## Description of the change

> This PR ingests `StatePersonAddressPeriod` as well as modifies the
logic for the `StatePersonHousingStatusPeriod` ingest view to ensure
that there are no overlapping periods. Note that for the `AddressPeriod`
entity, we filter out some entries in the `OCCUPANCY` table where the
person is not marked as unhoused/no address provided, but the address
field is null.

> [Successful test calc
DAG](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-07-28_07_25_21-11838422938666772726)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#44799

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ca34e36c7a2a7d5f82ba8a6bdc584298d90ca02f