[US_TN] Update movement filtering in IP view (apply filter earlier, change logic for final moves) (Recidiviz/recidiviz-data#45740)
## Description of the change

Certain movements in TN are treated as non-physical and filtered out if
they have one of a set of movement types and have location mismatches
with the surrounding movements. This change updates the filtering logic
in the IP view in two ways:
- Moves the filter to earlier in the view query. Currently, we're
checking for location mismatches *after* supervision-related movements
get filtered out of the movement data. This means that if someone is
released to supervision, gets transferred to a different supervision
location, and then is reincarcerated, the reincarceration movement may
be filtered out because when we check if the movement's location matches
with the previous movement, we aren't looking at the updated supervision
location from the supervision transfer movement.
- Changes the way we check for location mismatches for peoples' final
movements. Previously, a movement would never be filtered out if it was
a person's final movement, because that means we don't have a subsequent
movement to match locations against. With this change, that same logic
will apply for final movements with a NEWAD movement reason, but we will
now filter out non-NEWAD final movements if they meet the other
conditions (has an eligible movement type, "from" location doesn't match
the "to" location of the previous movement), AND if the person has
current site data in the Offender table. (If they don't have an assigned
site, that means their incarceration has ended, so we don't want to
create an open IP by filtering out their final discharge movement)

Here's an example to help understand why both parts of this change need
to take effect together. Say that someone's OffenderMovement data has:
- a release from prison A to supervision office X at t=1
- a transfer from supervision office X to supervision office Y at t=2
- a revocation from supervision office Y to county jail B at t=3

Under the old approach, the transfer from office X to Y wouldn't be
considered when checking for location mismatches. This means that the
person's final movement could get filtered out because 1) there's no
subsequent movement and it's not a NEWAD, and 2) the "from" location (Y)
doesn't match the previous "to" location (X). Under the new approach,
the final movement would not be filtered out because we consider Y to be
the previous location, so there's no mismatch.

_**FOR REVIEWERS:** Note that we previously filtered out these
non-physical movements by adding a flag and then filtering out movements
where the flag was TRUE. This change simplifies this slightly by
directly filtering out the movements with a WHERE, which means that all
the boolean logic is inverted (meaning that the ANDs are now ORs and all
the boolean operators have been flipped)._
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1733491cd6bf9ddf53a816a1ad34b56d0a3a91b4