[Cleanup] Remove downstream view usages of client_record's person_external_id col (Recidiviz/recidiviz-data#45707)
## Description of the change

We eventually want to delete the `person_external_id` column out of
`client_record` in favor of the more specifically named
`stable_person_external_id`. In order to do so, we need to remove all
downstream references to that column. For this PR, I commented out the
`person_external_id` column, replaced references to either `person_id`
or `stable_person_external_id`, where appropriate then re-added the
column (it's still referenced downstream so needs to stick around for
now).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#44536

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4bf41742a4ce4c442a09e24bdc451376117fcee4