[Resource Search] Resolve `engine_test.py` Failures and Enable CI Execution (Recidiviz/recidiviz-data#45738)
This PR resolves the non-executing and failing engine_test.py tests in
CI, as reported. Instead of further debugging the existing
implementation, I've reverted engine_test.py to the exact working
version from the prototype repository, with the exception of a "todo"
test case. This change also introduces a new dependency for
pytest-asyncio to correctly handle the asynchronous tests.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Helper Bot <helperbot@recidiviz.org>
GitOrigin-RevId: f1163f54cb0f9445303eceae697266524e797d8b