[US_AZ] Separate ineligible offense criteria into two (Recidiviz/recidiviz-data#45801)
## Description of the change

This PR now does two things, because I combined it with Recidiviz/recidiviz-data#45795

## Thing one
In the new version of the policy, there are some offenses that a person
cannot have _ever_ been convicted of, and some that they just can't be
actively serving a sentence for in order to be eligible. I think the
easiest way to assess those criteria is separately - this PR makes a new
criteria that checks if a person has ever had a conviction of the ones
that are permanently ineligible, and modifies the existing criteria to
only check for current offenses.

It also defines a new `_INELIGIBLE_OFFENSE_CRITERIA` variable with `AND`
logic that checks that both of the offense-related criteria are met.

## Thing two
After the most recent policy revision, the "stable housing" criteria
changed such that we can no longer verify it in ACIS using ORAS
assessment results. The relevant ORAS question is backward-looking (has
someone had stable housing for the last 12 months), but the eligibility
requirement is forward-looking (will someone have stable housing for the
next 6 months).

Because of this, we will not be doing any check of this criteria
automatically, and will instead surface it to users as a requirement to
check manually for each client. This PR implements that change by doing
the following:
- deletes the `not_homeless_in_release_plan` criteria builder
- removes the `not_homeless_in_release_plan` criteria from the
opportunity TES
- removes that criteria from the almost eligible criteria.
- removes that criteria from the `create_maybe_eligible_columns`
definition

The rest of the changes to make this appear as intended in the frontend
were already made in the admin panel and can be seen in staging.

## Eligibility changes
These two changes together increase eligibility by 29%!
<img width="1179" height="33" alt="Screenshot 2025-07-31 at 11 55 45 AM"
src="https://github.com/user-attachments/assets/73b7e40b-2c59-4612-ac74-b043114500d0"
/>

No validations are failing after this change either 🎉

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5c63f73efe1269f9ca6cb97a42036f5f7fac9126