Migrate TX contact standard criteria to state agnostic (Recidiviz/recidiviz-data#45816)
## Description of the change

A beefy change to migrate type-specific contacts criteria to use the
cool new state-agnostic function developed for NE! This does not affect
type-agnostic contacts which are still specific to TX. I ran a sandbox
to check for any changes, the only changes are some historical records
based on the change in logic to the overdue flag in the state_agnostic
function (which is expected and desired). I'll put the query for that
check in a comment.

| criteria_name             | presence_in_versions | row_count |
|---------------------------|----------------------|-----------|
| usTxHomeContactScheduled  | Only in Regular Staging | 747       |
| usTxHomeContactScheduled  | Both Versions        | 3         |
| usTxHomeContactScheduled  | Only in SPEPIN Staging | 1653      |
| usTxHomeContactUnscheduled| Only in Regular Staging | 370       |
| usTxHomeContactUnscheduled| Only in SPEPIN Staging | 182       |

Includes changes such as:
- Migrating us_tx_contact_cadence_spans to tasks_views instead of
analyst_data
- Updating the us_tx contact_compliance_builder to use the
state-agnostic compliance builder instead of containing it's own logic
- ensuring contact_cadence_spans and contact_events had appropriate
columns to support the state agnostic function
- update various columns in critically understaffed logic to account for
column names changing

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0789693010c5aed83af2c846e16bd228e2d2757f