[US_AZ] Fix 1.1/1.2/1.9 criteria for administrative supervision (Recidiviz/recidiviz-data#45937)
## Description of the change

We were previously combining criteria 1.1/1.2/1.9 into a single
`StateSpecificTaskCriteriaGroupBigQueryViewBuilder` that was buggy. This
separates the 1.1/1.9 and 1.2/1.9 checks into two distinct
`StateSpecificTaskCriteriaGroupBigQueryViewBuilder` so that this won't
be an issue.

This reduces eligibility by 21%
<img width="876" height="104" alt="Screenshot 2025-08-04 at 2 10 59 PM"
src="https://github.com/user-attachments/assets/bdaff7f6-699d-4100-9da2-0685c305a6c4"
/>

For example, here is the previous
`StateSpecificTaskCriteriaGroupBigQueryViewBuilder` (removed in this
PR):

```
# 1.9 Offenders with ineligible offenses are only eligible if they've been 15 months violation free
        StateSpecificTaskCriteriaGroupBigQueryViewBuilder(
            logic_type=TaskCriteriaGroupLogicType.OR,
            criteria_name="US_AZ_ANY_RISK_OR_INELIGIBLE_OFFENSES_BUT_15_MONTHS_VIOLATION_FREE",
            sub_criteria_list=[
                StateSpecificTaskCriteriaGroupBigQueryViewBuilder(
                    logic_type=TaskCriteriaGroupLogicType.OR,
                    criteria_name="US_AZ_ANY_RISK_OR_INELIG_OFFENSE",
                    sub_criteria_list=[
                        _INELIGIBLE_OFFENSE_CRITERIA,
                        _RISK_SCORE_CRITERIA,
                    ],
                    allowed_duplicate_reasons_keys=[],
                ),
                StateAgnosticTaskCriteriaGroupBigQueryViewBuilder(
                    logic_type=TaskCriteriaGroupLogicType.AND,
                    criteria_name="15_MONTHS_ON_SUPERVISION_VIOLATION_FREE",
                    sub_criteria_list=[
                        no_supervision_violation_within_15_months.VIEW_BUILDER,
                        on_supervision_at_least_15_months.VIEW_BUILDER,
                    ],
                    allowed_duplicate_reasons_keys=[],
                ),
            ],
            allowed_duplicate_reasons_keys=[],
        ),
```

In the case where a person has both a `MINIMUM` risk score and a felony
domestic violence conviction, and has only been on supervision for 1
month, the following would be true:
- They have an eligible risk score [`_RISK_SCORE_CRITERIA = True`
- They have an ineligible offense [`_INELIGIBLE_OFFENSE_CRITERIA =
False`]
- They have not been on supervision violation-free for 15 months
[`15_MONTHS_ON_SUPERVISION_VIOLATION_FREE = False`]

Taking all of these together with the grouping logic, we would have:
`((True or False) or False)`, which would make the condition `True` as a
whole.

However, when a person has a felony DV conviction and only been on
supervision for one month, they cannot be eligible for this opportunity.
The eligible risk score in this case was obfuscating the ineligible
offense. Separating these two checks protects us against that.
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Damini Sharma <damini@recidiviz.org>
GitOrigin-RevId: a915fbb601455349e9f9d850d5584b20aee0862b