[Ingest] Enforce that StateStaff emails are unqiue (case-insensitive) for all states going forward (Recidiviz/recidiviz-data#45916)
## Description of the change

Adds some new features to our `UnqiueConstraint` checks that allow us to
add a case-insensitive uniqueness constraint on `StateStaff.email`:
1) Adds a state-specific exemption mechanism to unique constraints so
that we can turn on a uniqueness constraint for all states going forward
and prevent backslide
2) Updates our uniqueness constraint checks to apply to fields that may
have null values: we allow multiple rows with a null value but do not
otherwise allow multiple rows with any other nonnull value.
3) Updates UniqueConstraint to allow us to apply a transform before we
check for uniqueness. This allows us to do a case-insensitive check for
emails.

[This](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-08-01_16_19_57-14094332300695179318;logsSeverity=INFO;graphView=0?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))&authuser=0&inv=1&invt=Ab4WlA)
is an example pipeline that failed (I had temporarily excluded IA from
the exemption) and
[this](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-08-01_16_23_58-15641934290740747344;logsSeverity=INFO;graphView=0?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))&authuser=0&inv=1&invt=Ab4WlA)
is an example pipeline that passed even though there are duplicates
(i.e. the exemption is working).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes
[Recidiviz/recidiviz-data#29073](https://github.com/Recidiviz/recidiviz-data/issues/29073)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 765a782a5677fcb0ed79fad55da8b7295e938aff